release/0.3.0: change Func and Result.Processed to int64 and change assignment.Decision.Slot to uint64 and other minor changes

+++ job.go
@@ -8,13 +8,15 @@ import (
 // Func is one attempt at the work. It reports how many items it processed -
 // return 0 when the work is not item-oriented - and an error if the attempt
 // failed.
+// The count is expected to be non-negative; the runner preserves it as returned
+// rather than validating or clamping it.
 //
 // A Func must respect ctx: the runner bounds it with the configured timeout,
 // and whoever owns the process cancels it on shutdown. The bound is
 // cooperative, as every deadline in Go is: it cancels the context, it does not
 // interrupt the running function. A Func must return when its context is done,
 // and must wait for its own goroutines before it returns.
-type Func func(ctx context.Context) (int, error)
+type Func func(ctx context.Context) (int64, error)
 
 // Middleware wraps a Func to add behaviour around it - recovery, context
 // enrichment, extra logging. It is classic Go composition: a Middleware
